se acomodan funciones, se resuelven conflictos
