+++ main/entities/utils.py
@@ -1,0 +1,3 @@
+def getById(id,arr):
+    filtered = list(filter(lambda value : value[0]==id,arr))
+    return filtered[0] if (filtered != None and len(filtered)>0 )else -1
